+++ tests/CfdiUtilsTests/Certificado/SerialNumberTest.php
@@ -12,13 +12,12 @@ class SerialNumberTest extends TestCase
         $expectedDecimal = '292233162870206001759766198425879490508935868472';
         $expectedAscii = '30001000000300023708';
         $serial = new SerialNumber($input);
+        $this->assertSame($input, $serial->getHexadecimal());
         $this->assertSame($expectedDecimal, $serial->asDecimal());
         $this->assertSame($expectedAscii, $serial->asAscii());
     }
 
     /**
-     * This test extends the base class to expose the visibility of serialHexToAscii method.
-     *
      * @param string $input
      * @param string $expected
      * @testWith ["3330303031303030303030333030303233373038", "30001000000300023708"]
@@ -32,8 +31,6 @@ class SerialNumberTest extends TestCase
     }
 
     /**
-     * This test extends the base class to expose the visibility of serialHexToAscii method.
-     *
      * @param string $input
      * @param string $expected
      * @testWith ["0x3330303031303030303030333030303233373038", "30001000000300023708"]
@@ -44,5 +41,17 @@ class SerialNumberTest extends TestCase
         $serial = new SerialNumber('');
         $serial->loadDecimal($input);
         $this->assertSame($expected, $serial->asAscii());
+    }
+
+    /**
+     * @param string $input
+     * @param string $expected
+     * @testWith ["30001000000300023708", "3330303031303030303030333030303233373038"]
+     */
+    public function testLoadAscii(string $input, string $expected)
+    {
+        $serial = new SerialNumber('');
+        $serial->loadAscii($input);
+        $this->assertSame($expected, $serial->getHexadecimal());
     }
 }